% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_w_histograms.R
\name{explore_w_histograms}
\alias{explore_w_histograms}
\title{take a csv file, a string of numerical column names
and returns a png file of histogram(s)}
\usage{
explore_w_histograms(file, column_names)
}
\arguments{
\item{file}{csv file path}

\item{column_names}{a string of numerical column names, seperate with comma}
}
\value{
.png file(s) of histogram(s)
}
\description{
take a csv file, a string of numerical column names
and returns a png file of histogram(s)
}
\examples{
explore_w_histograms(file = 'toy_dataset.csv', column_names = "volumn,date"])
}
