% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_w_histograms.R
\name{is_numeric}
\alias{is_numeric}
\title{Helper function used to take a dataframe, a column name
Returns True if the column is numerical, False otherwise}
\usage{
is_numeric(df, column)
}
\arguments{
\item{df}{data.frame}

\item{column}{str, the name of the column}
}
\value{
logical TRUE or FALSE
}
\description{
Helper function used to take a dataframe, a column name
Returns True if the column is numerical, False otherwise
}
