% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_overview.R, R/sample_data.R,
%   R/word_bubble.R
\name{load_file}
\alias{load_file}
\title{Helper function to read the file and return a dataframe.}
\usage{
load_file(file, sheet_name = NULL)

load_file(file, sheet_name = NULL)

load_file(file, sheet_name = NULL)
}
\arguments{
\item{file}{the name of the file, including the path and filetype extension}

\item{sheet_name}{if passing an excel file, the name of the sheet to analyze (default = NULL)}
}
\value{
dataframe

dataframe

dataframe
}
\description{
Checks if file is type .csv or excel. If not, return an error.

Checks if file is type .csv or excel. If not, return an error.

Checks if file is type .csv or excel. If not, return an error.
}
