% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_w_histograms.R
\name{make_histogram}
\alias{make_histogram}
\title{Helper function used to take a dataframe, a numerical column name,}
\usage{
make_histogram(df, column)
}
\arguments{
\item{df}{data.frame}

\item{column}{str, the name of the column}
}
\value{
a histogram of the column
}
\description{
Helper function used to take a dataframe, a numerical column name,
}
