% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_bubble.R
\name{make_plot}
\alias{make_plot}
\title{Helper function to process the corpus into a form that word cloud can use.}
\usage{
make_plot(freq, dir, max, height, width)
}
\arguments{
\item{freq}{frequency of the words from the output of the process corpus function}

\item{dir}{the directory where the file should be saved, default = ''}

\item{max}{number of words in the word bubble, default = Inf words}

\item{height}{the height of the outputted image}

\item{width}{the width of the outputted image}
}
\value{

}
\description{
Helper function to process the corpus into a form that word cloud can use.
}
