% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_overview.R
\name{missing_data_overview}
\alias{missing_data_overview}
\title{Return a heatmap showing the missing values in the file.}
\usage{
missing_data_overview(file, sheet_name = NULL, dir = "")
}
\arguments{
\item{file}{the name of the file, including the filetype extension}

\item{sheet_name}{if passing an excel file, the name of the sheet to analyze, by default NULL}

\item{dir}{the directory where the file should be saved, by default ''}
}
\value{
.png file heatmap of missing values, as a .png file
}
\description{
Return a heatmap showing the missing values in the file.
}
\examples{
missing_data_overview("customers.xlsx", sheet_name='2019', dir='report')
}
