% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_bubble.R
\name{process_corpus}
\alias{process_corpus}
\title{Helper function to read the file and return a dataframe.}
\usage{
process_corpus(df)
}
\arguments{
\item{df}{dataframe of the text column}
}
\value{
formated corpus with frequencys of each word
}
\description{
Helper function to read the file and return a dataframe.
}
