% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_w_histograms.R
\name{read_file}
\alias{read_file}
\title{Helper function used to read the file and return a dataframe.
Checks if file type is a .csv or excel. If not, it prints a message and
returns NA}
\usage{
read_file(file, sheet_name = NULL)
}
\arguments{
\item{file}{str,the path of the file, including the filetype extension}

\item{sheet_name}{int, default NULL
if passing an excel file, the name of the sheet to analyze}
}
\value{
data.frame
}
\description{
Helper function used to read the file and return a dataframe.
Checks if file type is a .csv or excel. If not, it prints a message and
returns NA
}
