% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{summarize_data}
\alias{summarize_data}
\title{Helper function used to create summary data table showing column names in rows,
an example record, column data types and summary statistics for different data types}
\usage{
summarize_data(df)
}
\arguments{
\item{df}{the dataframe object to analyze}
}
\value{
dataframe
}
\description{
Helper function used to create summary data table showing column names in rows,
an example record, column data types and summary statistics for different data types
}
