% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_bubble.R
\name{word_bubble}
\alias{word_bubble}
\title{Return an image of a word bubble of qualitative responses (text)
from a column(s) from a spreadsheet.}
\usage{
word_bubble(
  file,
  sheet_name = 0,
  dir = "",
  column = "",
  max = 50,
  height = 8,
  width = 8
)
}
\arguments{
\item{file}{the name of the file, including the filetype extension}

\item{sheet_name}{if passing an excel file, the name of the sheet to analyze, default =  0}

\item{dir}{the directory where the file should be saved, default = ''}

\item{column}{A string or vector of the columns header that are to be included in the word bubble, default = ''}

\item{max}{number of words in the word bubble, default = 50 words}

\item{height}{the height of the outputted image, default = 8 inches}

\item{width}{the width of the outputted image, default = 8 inches}
}
\value{
returns an image of a word bubble by specified width and height and includes max number of words.
}
\description{
Return an image of a word bubble of qualitative responses (text)
from a column(s) from a spreadsheet.
}
\examples{
sample_data(customers.xlsx, sheet_name='2019', dir='report', column='review', max=50, height=7, width=7)
}
