% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_w_histograms.R
\name{explore_w_histograms}
\alias{explore_w_histograms}
\title{take a csv file path, a sheet name (default NULL),
a list of numerical column names
and returns a png file of histogram(s)}
\usage{
explore_w_histograms(file, columns_list, sheet_name = NULL, dir = "")
}
\arguments{
\item{file}{str, the path of the file, including the filetype extension}

\item{columns_list}{a list of numerical column names as string}

\item{sheet_name}{int, default NULL
if passing an excel file, the name of the sheet to analyze}

\item{dir}{str, default ''
the directory where the file should be saved,}
}
\value{
printed messages
}
\description{
take a csv file path, a sheet name (default NULL),
a list of numerical column names
and returns a png file of histogram(s)
}
