% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_overview.R
\name{make_plot}
\alias{make_plot}
\title{Helper function used to create heatmap showing missing values in a dataframe.}
\usage{
make_plot(df)
}
\arguments{
\item{df}{the dataframe object to analyze}
}
\value{
ggplot heatmap
}
\description{
Helper function used to create heatmap showing missing values in a dataframe.
}
